requirements.txt
django==3.0.2
pillow==7.0.0
